backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.6.0
pydantic[email]
boto3==1.35.0
psycopg2-binary==2.9.10
